Fix #279.
